src/tag-normalization/requirements.txt
# Tag Normalization & Inference Engine Dependencies
# Python 3.8+

# Core dependencies (optional but recommended for semantic analysis)
scikit-learn>=1.0.0          # TF-IDF vectorizer and cosine similarity for semantic analysis
numpy>=1.21.0               # Numerical computing (required by sklearn)

# Reporting & Visualization
tabulate>=0.9.0             # Pretty-print tables for reports

# Optional: Advanced NLP (uncomment if needed)
# spacy>=3.0.0              # Advanced NLP for Portuguese text analysis
# nltk>=3.6.0               # Natural Language Toolkit
# transformers>=4.0.0       # Hugging Face transformers for embeddings

# Development & Testing
pytest>=6.2.0               # Unit testing framework
pyyaml>=5.4.0               # YAML parsing (already used in script)
